+++ Assets/Melody Render Pipeline/ShaderLibrary/ScreenSpaceTrace.hlsl
@@ -1,66 +1,66 @@
 #ifndef MELODY_SCREEN_SPACE_TRACE_INCLUDED
 #define MELODY_SCREEN_SPACE_TRACE_INCLUDED
 
-inline half GetScreenFadeBord(half2 pos, half value) {
-    half borderDist = min(1 - max(pos.x, pos.y), min(pos.x, pos.y));
+inline float GetScreenFadeBord(float2 pos, float value) {
+    float borderDist = min(1 - max(pos.x, pos.y), min(pos.x, pos.y));
     return saturate(borderDist > value ? 1 : borderDist / value);
 }
 
-inline half3 ReconstructCSPosition(half4 _MainTex_TexelSize, half4 _ProjInfo, half2 S, half z) {
-    half linEyeZ = -LinearEyeDepth(z, _ZBufferParams);
-    return half3((((S.xy * _MainTex_TexelSize.zw)) * _ProjInfo.xy + _ProjInfo.zw) * linEyeZ, linEyeZ);
+inline float3 ReconstructCSPosition(float4 _MainTex_TexelSize, float4 _ProjInfo, float2 S, float z) {
+    float linEyeZ = -LinearEyeDepth(z, _ZBufferParams);
+    return float3((((S.xy * _MainTex_TexelSize.zw)) * _ProjInfo.xy + _ProjInfo.zw) * linEyeZ, linEyeZ);
 }
 
-inline half3 GetPosition(TEXTURE2D(depth), half4 _MainTex_TexelSize, half4 _ProjInfo, half2 ssP) {
-    half3 P;
+inline float3 GetPosition(TEXTURE2D(depth), float4 _MainTex_TexelSize, float4 _ProjInfo, float2 ssP) {
+    float3 P;
     P.z = SAMPLE_DEPTH_TEXTURE(depth, sampler_point_clamp, ssP.xy).r;
-    P = ReconstructCSPosition(_MainTex_TexelSize, _ProjInfo, half2(ssP), P.z);
+    P = ReconstructCSPosition(_MainTex_TexelSize, _ProjInfo, float2(ssP), P.z);
     return P;
 }
 
 //2D linear trace
-inline half distanceSquared(half2 A, half2 B) {
+inline float distanceSquared(float2 A, float2 B) {
     A -= B;
     return dot(A, A);
 }
 
-inline half distanceSquared(half3 A, half3 B) {
+inline float distanceSquared(float3 A, float3 B) {
     A -= B;
     return dot(A, A);
 }
 
-void swap(inout half v0, inout half v1) {
-    half temp = v0;
+void swap(inout float v0, inout float v1) {
+    float temp = v0;
     v0 = v1;
     v1 = temp;
 }
 
-bool intersectsDepthBuffer(half rayZMin, half rayZMax, half sceneZ, half layerThickness) {
+bool intersectsDepthBuffer(float rayZMin, float rayZMax, float sceneZ, float layerThickness) {
     return (rayZMax >= sceneZ - layerThickness) && (rayZMin <= sceneZ);
 }
 
 void rayIterations(TEXTURE2D(forntDepth), 
     in bool traceBehind_Old, 
     in bool traceBehind, 
-    inout half2 P, 
-    inout half stepDirection, 
-    inout half end, 
+    inout float2 P, 
+    inout float stepDirection, 
+    inout float end, 
     inout int stepCount, 
     inout int maxSteps, 
     inout bool intersecting,
-    inout half sceneZ, 
-    inout half2 dP, 
-    inout half3 Q, 
-    inout half3 dQ, 
-    inout half k, 
-    inout half dk,
-    inout half rayZMin, 
-    inout half rayZMax, 
-    inout half prevZMaxEstimate, 
+    inout float sceneZ, 
+    inout float2 dP, 
+    inout float3 Q, 
+    inout float3 dQ, 
+    inout float k, 
+    inout float dk,
+    inout float rayZMin, 
+    inout float rayZMax, 
+    inout float prevZMaxEstimate, 
     inout bool permute, 
-    inout half2 hitPixel,
-    half2 invSize, 
-    inout half layerThickness) {
+    inout float2 hitPixel,
+    float2 invSize, 
+    inout float layerThickness) {
     bool stop = intersecting;
     for (; (P.x * stepDirection) <= end && stepCount < maxSteps && !stop; P += dP, Q.z += dQ.z, k += dk, stepCount += 1)
     {
@@ -73,7 +73,7 @@ void rayIterations(TEXTURE2D(forntDepth),
         }
 
         hitPixel = permute ? P.yx : P;
-        sceneZ = SAMPLE_TEXTURE2D_LOD(forntDepth, sampler_point_clamp, half2(hitPixel * invSize), 0).r;
+        sceneZ = SAMPLE_TEXTURE2D_LOD(forntDepth, sampler_point_clamp, float2(hitPixel * invSize), 0).r;
         sceneZ = -LinearEyeDepth(sceneZ, _ZBufferParams);
         bool isBehind = (rayZMin <= sceneZ);
 
@@ -90,49 +90,49 @@ void rayIterations(TEXTURE2D(forntDepth),
 }
 
 bool Linear2D_Trace(TEXTURE2D(forntDepth),
-    half3 csOrigin,
-    half3 csDirection,
-    half4x4 projectMatrix,
-    half2 csZBufferSize,
-    half jitter,
+    float3 csOrigin,
+    float3 csDirection,
+    float4x4 projectMatrix,
+    float2 csZBufferSize,
+    float jitter,
     int maxSteps,
-    half layerThickness,
-    half traceDistance,
-    in out half2 hitPixel,
+    float layerThickness,
+    float traceDistance,
+    in out float2 hitPixel,
     int stepSize,
     bool traceBehind,
-    in out half3 csHitPoint,
-    in out half stepCount) {
+    in out float3 csHitPoint,
+    in out float stepCount) {
 
-    half2 invSize = half2(1 / csZBufferSize.x, 1 / csZBufferSize.y);
-    hitPixel = half2(-1, -1);
+    float2 invSize = float2(1 / csZBufferSize.x, 1 / csZBufferSize.y);
+    hitPixel = float2(-1, -1);
 
-    half nearPlaneZ = -0.01;
-    half rayLength = ((csOrigin.z + csDirection.z * traceDistance) > nearPlaneZ) ? ((nearPlaneZ - csOrigin.z) / csDirection.z) : traceDistance;
-    half3 csEndPoint = csDirection * rayLength + csOrigin;
-    half4 H0 = mul(projectMatrix, half4(csOrigin, 1));
-    half4 H1 = mul(projectMatrix, half4(csEndPoint, 1));
-    half k0 = 1 / H0.w;
-    half k1 = 1 / H1.w;
-    half2 P0 = H0.xy * k0;
-    half2 P1 = H1.xy * k1;
-    half3 Q0 = csOrigin * k0;
-    half3 Q1 = csEndPoint * k1;
+    float nearPlaneZ = -0.01;
+    float rayLength = ((csOrigin.z + csDirection.z * traceDistance) > nearPlaneZ) ? ((nearPlaneZ - csOrigin.z) / csDirection.z) : traceDistance;
+    float3 csEndPoint = csDirection * rayLength + csOrigin;
+    float4 H0 = mul(projectMatrix, float4(csOrigin, 1));
+    float4 H1 = mul(projectMatrix, float4(csEndPoint, 1));
+    float k0 = 1 / H0.w;
+    float k1 = 1 / H1.w;
+    float2 P0 = H0.xy * k0;
+    float2 P1 = H1.xy * k1;
+    float3 Q0 = csOrigin * k0;
+    float3 Q1 = csEndPoint * k1;
 
-    half yMax = csZBufferSize.y - 0.5;
-    half yMin = 0.5;
-    half xMax = csZBufferSize.x - 0.5;
-    half xMin = 0.5;
-    half alpha = 0;
+    float yMax = csZBufferSize.y - 0.5;
+    float yMin = 0.5;
+    float xMax = csZBufferSize.x - 0.5;
+    float xMin = 0.5;
+    float alpha = 0;
 
     if (P1.y > yMax || P1.y < yMin) {
-        half yClip = (P1.y > yMax) ? yMax : yMin;
-        half yAlpha = (P1.y - yClip) / (P1.y - P0.y);
+        float yClip = (P1.y > yMax) ? yMax : yMin;
+        float yAlpha = (P1.y - yClip) / (P1.y - P0.y);
         alpha = yAlpha;
     }
     if (P1.x > xMax || P1.x < xMin) {
-        half xClip = (P1.x > xMax) ? xMax : xMin;
-        half xAlpha = (P1.x - xClip) / (P1.x - P0.x);
+        float xClip = (P1.x > xMax) ? xMax : xMin;
+        float xAlpha = (P1.x - xClip) / (P1.x - P0.x);
         alpha = max(alpha, xAlpha);
     }
 
@@ -140,8 +140,8 @@ bool Linear2D_Trace(TEXTURE2D(forntDepth),
     k1 = lerp(k1, k0, alpha);
     Q1 = lerp(Q1, Q0, alpha);
 
-    P1 = (distanceSquared(P0, P1) < 0.0001) ? P0 + half2(0.01, 0.01) : P1;
-    half2 delta = P1 - P0;
+    P1 = (distanceSquared(P0, P1) < 0.0001) ? P0 + float2(0.01, 0.01) : P1;
+    float2 delta = P1 - P0;
     bool permute = false;
 
     if (abs(delta.x) < abs(delta.y)) {
@@ -151,11 +151,11 @@ bool Linear2D_Trace(TEXTURE2D(forntDepth),
         P0 = P0.yx;
     }
 
-    half stepDirection = sign(delta.x);
-    half invdx = stepDirection / delta.x;
-    half2 dP = half2(stepDirection, invdx * delta.y);
-    half3 dQ = (Q1 - Q0) * invdx;
-    half dk = (k1 - k0) * invdx;
+    float stepDirection = sign(delta.x);
+    float invdx = stepDirection / delta.x;
+    float2 dP = float2(stepDirection, invdx * delta.y);
+    float3 dQ = (Q1 - Q0) * invdx;
+    float dk = (k1 - k0) * invdx;
 
     dP *= stepSize;
     dQ *= stepSize;
@@ -164,15 +164,15 @@ bool Linear2D_Trace(TEXTURE2D(forntDepth),
     Q0 += dQ * jitter;
     k0 += dk * jitter;
 
-    half3 Q = Q0;
-    half k = k0;
-    half prevZMaxEstimate = csOrigin.z;
+    float3 Q = Q0;
+    float k = k0;
+    float prevZMaxEstimate = csOrigin.z;
     stepCount = 0;
-    half rayZMax = prevZMaxEstimate, rayZMin = prevZMaxEstimate;
-    half sceneZ = 100000;
-    half end = P1.x * stepDirection;
+    float rayZMax = prevZMaxEstimate, rayZMin = prevZMaxEstimate;
+    float sceneZ = 100000;
+    float end = P1.x * stepDirection;
     bool intersecting = intersectsDepthBuffer(rayZMin, rayZMax, sceneZ, layerThickness);
-    half2 P = P0;
+    float2 P = P0;
     int originalStepCount = 0;
 
     bool traceBehind_Old = true;
