+++ Assets/Melody Render Pipeline/Shaders/StochasticSSRPass.hlsl
@@ -239,8 +239,8 @@ void LinearTraceMultiSPP(Varyings input, out float4 SSRColor_PDF : SV_TARGET0, o
 	Out_UV /= _SSR_NumRays;
 	Out_PDF /= _SSR_NumRays;
 
-	SSRColor_PDF = half4(Out_Color.rgb, Out_PDF);
-	Mask_Depth_HitUV = half4(Square(Out_Mask), Out_RayDepth, Out_UV);
+	SSRColor_PDF = float4(Out_Color.rgb, Out_PDF);
+	Mask_Depth_HitUV = float4(Square(Out_Mask), Out_RayDepth, Out_UV);
 }
